Set source enconding and complier plugin version
The `mvn test` command prints two warnings.

1. Compiler plugin without version:

```
[WARNING] 
[WARNING] Some problems were encountered while building the effective model for org.brexhq:hello_world_java:jar:1.0-SNAPSHOT
[WARNING] 'build.plugins.plugin.version' for org.apache.maven.plugins:maven-compiler-plugin is missing. @ line 12, column 21
[WARNING] 
[WARNING] It is highly recommended to fix these problems because they threaten the stability of your build.
[WARNING] 
[WARNING] For this reason, future Maven versions might no longer support building such malformed projects.
[WARNING] 
```

2.  Build is platform dependent:
```
[WARNING] Using platform encoding (UTF-8 actually) to copy filtered resources, i.e. build is platform dependent!
```

Both are minor problems but also easy to fix.
The first may (very unlikely) cause errors/inconsistencies whenever a new version of the the plugin is released. Setting the version guarantees consistency in the future.
The later is more likely to happen given the project reads files. The output of the build on Windows (encoding cp-1252) might be different of the output on Linux (UTF-8 probably).

This PR sets the configurations to avoid the WARNING and its possible side-effects.